[FP-22] Refactor: useCardStatus.tsx 커스텀 훅을 CardStatus.ts 헬퍼 함수로 변경
